refactor: extract graphql request hook into separate file
